cleaned up component folders for better conventions
