+++ client/src/pages/Login.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import LoginForm from '../components/LoginForm/LoginForm';
+import LoginForm from '../components/LoginForm';
 
 function Login(props) {
   return (
